fix(contracts): move domain types from tests/ to service modules (#5)
Production code was importing enums and dataclasses from tests/contracts/,
which would fail if tests are excluded from deployment.

- Create services/aggregator_service/domain.py with ServerTransportType,
  ServerStatus, RoutingStrategy
- Create services/marketplace_service/domain.py with RegistrySource,
  InstallStatus, UpdateChannel, PackageSpec, InstallResult, UpdateInfo
- Update all 12 service files to import from .domain instead
- Reverse test contract dependency: tests now import FROM domain modules
- SearchResult remains test-only with try/except fallback
