+++ services/aggregator_service/aggregator_service.py
@@ -10,10 +10,7 @@ from datetime import datetime, timezone
 from typing import Any, Dict, List, Optional
 import logging
 
-from tests.contracts.aggregator.data_contract import (
-    ServerTransportType,
-    ServerStatus,
-)
+from .domain import ServerTransportType, ServerStatus
 
 from .server_registry import ServerRegistry
 from .session_manager import SessionManager
